requirements.txt
aerich==0.9.0
aiomysql==0.2.0
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8
certifi==2025.4.26
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
coverage==7.8.2
dictdiffer==0.9.0
distlib==0.3.9
dnspython==2.7.0
email-validator==2.2.0
faker==37.3.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
loguru==0.7.3
mako==1.3.10
markupsafe==3.0.2
mypy==1.16.0
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==4.2.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pymysql==1.1.1
pypika-tortoise==0.6.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
pytest-env==1.1.5
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
ruff==0.11.13
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tomlkit==0.13.3
tortoise-orm==0.25.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.3
virtualenv==20.31.2
win32-setctime==1.2.0
